master

+++ advanced_features_and_security/LibraryProject/bookshelf/views.py
@@ -1,7 +1,7 @@
-from django.shortcuts import render, get_object_or_404
-from django.http import HttpResponseForbidden
+from django.shortcuts import render, get_object_or_404, redirect
 from django.contrib.auth.decorators import login_required, permission_required
 from .models import Book
+from .forms import BookForm
 
 @login_required
 @permission_required('bookshelf.can_view', raise_exception=True)
@@ -12,18 +12,34 @@ def book_list(request):
 @login_required
 @permission_required('bookshelf.can_create', raise_exception=True)
 def book_create(request):
-    # Logic to handle form submission
-    return render(request, 'bookshelf/book_form.html')
+    if request.method == 'POST':
+        form = BookForm(request.POST)
+        if form.is_valid():
+            form.save()
+            return redirect('book_list')
+    else:
+        form = BookForm()
+    return render(request, 'bookshelf/book_form.html', {'form': form})
 
 @login_required
 @permission_required('bookshelf.can_edit', raise_exception=True)
 def book_edit(request, book_id):
     book = get_object_or_404(Book, id=book_id)
-    return render(request, 'bookshelf/book_edit.html', {'book': book})
+    if request.method == 'POST':
+        form = BookForm(request.POST, instance=book)
+        if form.is_valid():
+            form.save()
+            return redirect('book_list')
+    else:
+        form = BookForm(instance=book)
+    return render(request, 'bookshelf/book_edit.html', {'form': form})
 
 @login_required
 @permission_required('bookshelf.can_delete', raise_exception=True)
 def book_delete(request, book_id):
     book = get_object_or_404(Book, id=book_id)
-    book.delete()
-    return redirect('book_list')
+    if request.method == 'POST':
+        book.delete()
+        return redirect('book_list')
+    return render(request, 'bookshelf/book_confirm_delete.html', {'book': book})
+
